backend/requirements.txt
djongo>=1.2
asgiref==3.3.1
Django==3.0.5
djangorestframework==3.12.2
psycopg2-binary==2.8.6
pymongo==3.11.3
pytz==2021.1
sqlparse==0.2.4
